+++ src/main/kotlin/be/ugent/webdevelopment/backend/geocode/controllers/wrappers/CurrentLocationWrapper.kt
@@ -1,6 +1,6 @@
 package be.ugent.webdevelopment.backend.geocode.controllers.wrappers
 
 class CurrentLocationWrapper(
-        var latitude: Int,
-        var longitude: Int
+        var latitude: Double,
+        var longitude: Double
 ): Wrapper()
